driver/gpu: correct ptr type errors

+++ drivers/gpu/mediatek/gpu_rgx/m1.13ED5776728/services/system/rogue/rgx_mtk/mt6765/mtk_mfgsys.c
@@ -333,34 +333,34 @@ static int MTKInitHWAPM(void)
 
 	if (gpu_debug_enable) {
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x504), mfg_readl(g_pvRegsKM + 0x504)));
+		(unsigned int)(g_pvRegsKM + 0x504), mfg_readl(g_pvRegsKM + 0x504)));
 
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x508), mfg_readl(g_pvRegsKM + 0x508)));
+		(unsigned int)(g_pvRegsKM + 0x508), mfg_readl(g_pvRegsKM + 0x508)));
 
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x50c), mfg_readl(g_pvRegsKM + 0x50c)));
+		(unsigned int)(g_pvRegsKM + 0x50c), mfg_readl(g_pvRegsKM + 0x50c)));
 
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x510), mfg_readl(g_pvRegsKM + 0x510)));
+		(unsigned int)(g_pvRegsKM + 0x510), mfg_readl(g_pvRegsKM + 0x510)));
 
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x514), mfg_readl(g_pvRegsKM + 0x514)));
+		(unsigned int)(g_pvRegsKM + 0x514), mfg_readl(g_pvRegsKM + 0x514)));
 
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x518), mfg_readl(g_pvRegsKM + 0x518)));
+		(unsigned int)(g_pvRegsKM + 0x518), mfg_readl(g_pvRegsKM + 0x518)));
 
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x51c), mfg_readl(g_pvRegsKM + 0x51c)));
+		(unsigned int)(g_pvRegsKM + 0x51c), mfg_readl(g_pvRegsKM + 0x51c)));
 
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x520), mfg_readl(g_pvRegsKM + 0x520)));
+		(unsigned int)(g_pvRegsKM + 0x520), mfg_readl(g_pvRegsKM + 0x520)));
 
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x524), mfg_readl(g_pvRegsKM + 0x524)));
+		(unsigned int)(g_pvRegsKM + 0x524), mfg_readl(g_pvRegsKM + 0x524)));
 
 		PVR_DPF((PVR_DBG_ERROR, "HWAPM: *g_pvRegsKM+0x%x = 0x%x",
-		(g_pvRegsKM + 0x528), mfg_readl(g_pvRegsKM + 0x528)));
+		(unsigned int)(g_pvRegsKM + 0x528), mfg_readl(g_pvRegsKM + 0x528)));
 
 		PVR_DPF((PVR_DBG_ERROR, "LV1 *g_pvRegsKM = 0x%x",
 		mfg_readl(g_pvRegsKM+0x01c)));
